+++ src/viz/__tests__/styles/ColorContinuousStyle.spec.ts
@@ -29,8 +29,8 @@ jest.mock('../../source/DatasetSource', () => ({
 
 const styledLayer = {
   getId: () => uuidv4(),
-  getMapInstance: () => ({} as Deck),
-  source: new DatasetSource('table')
+  getMap: () => ({} as Deck),
+  getSource: () => new DatasetSource('table')
 };
 
 describe('ColorContinuousStyle', () => {
@@ -99,7 +99,7 @@ describe('ColorContinuousStyle', () => {
         property: 'strokeColor'
       });
       const styleLayerPoint = { ...styledLayer };
-      styleLayerPoint.source.getMetadata = jest.fn().mockImplementation(() => {
+      styleLayerPoint.getSource().getMetadata = jest.fn().mockImplementation(() => {
         return {
           geometryType: 'Point',
           stats: [stats]