+++ src/viz/__tests__/styles/SizeContinuousStyle.spec.ts
@@ -28,8 +28,8 @@ jest.mock('../../source/DatasetSource', () => ({
 
 const styledLayer = {
   getId: () => uuidv4(),
-  getMapInstance: () => ({} as Deck),
-  source: new DatasetSource('table')
+  getMap: () => ({} as Deck),
+  getSource: () => new DatasetSource('table')
 };
 
 describe('SizeContinuousStyle', () => {