+++ src/viz/style/helpers/size-categories-style.ts
@@ -50,9 +50,9 @@ export function sizeCategoriesStyle(
   options: Partial<SizeCategoriesOptionsStyle> = {}
 ) {
   const evalFN = async (layer: StyledLayer) => {
-    const meta = layer.source.getMetadata();
+    const meta = layer.getSource().getMetadata();
 
-    if (layer.source.isEmpty()) {
+    if (layer.getSource().isEmpty()) {
       return {};
     }
 
@@ -69,7 +69,7 @@ export function sizeCategoriesStyle(
     layer: StyledLayer,
     legendWidgetOptions: LegendWidgetOptions = { config: {} }
   ): Promise<LegendProperties[]> => {
-    const meta = layer.source.getMetadata();
+    const meta = layer.getSource().getMetadata();
 
     if (!meta.geometryType) {
       return [];