+++ src/viz/style/helpers/size-continuous-style.ts
@@ -40,9 +40,9 @@ export function sizeContinuousStyle(
   options: Partial<SizeContinuousOptionsStyle> = {}
 ) {
   const evalFN = async (layer: StyledLayer) => {
-    const meta = layer.source.getMetadata();
+    const meta = layer.getSource().getMetadata();
 
-    if (layer.source.isEmpty()) {
+    if (layer.getSource().isEmpty()) {
       return {};
     }
 
@@ -79,7 +79,7 @@ export function sizeContinuousStyle(
     layer: StyledLayer,
     legendWidgetOptions: LegendWidgetOptions = { config: {} }
   ): Promise<LegendProperties[]> => {
-    const meta = layer.source.getMetadata();
+    const meta = layer.getSource().getMetadata();
 
     if (!meta.geometryType) {
       return [];